requirement.txt
# Install PyTorch with CUDA 11.8 support using:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

torch
torchvision
torchaudio
hydra-core
ray
flwr


